feat: general store
