--- storage/src/general_store/stub_general_store.rs
@@ -1,1 +1,0 @@
-
